rename file useImagePlayer.ts to usePlaylist.ts
